requirements.txt
pillow==9.5.0
cs50==9.4.0
flask_session==0.8.0
opencv-python==4.11.0.86
mediapipe==0.10.21
twilio==9.8.0